refactor (ArticleEndpoints): modified the update and create endpoints
