--- ApiRestMinimal/DTOs/ArticleDTOs.cs
@@ -1,9 +1,0 @@
-﻿namespace ApiRestMinimal.DTOs;
-
-public class ArticleDTOs
-{
-    public Guid Id { get; set; }
-    public string Title { get; set; }
-    public string Content { get; set; }
-    public Guid CategoryId { get; set; }
-}
